Added some timings to prove whether the threads were running at the same time.  They are not right now.  Created 3 Spring managed instances of the AsynchronousBean class.
Added an aop designation for AspectJ in the Spring config.  Also a task
executor.  The threads are not running at the same time yet so there are
still issue to work out.

The AspectJ compiler is upgraded for Maven and the libraries.  It is using
Java 1.7 but is still not quite right for compilation.